JGit bug workaround
Press uses JGit for syncing notes on Android, but it has an annoying
bug where commits without a parent get dropped after a rebase, which 
is always going to be true for the first commit. As a workaround, 
Press adds a dummy commit that sacrifices itself to protect an actual
commit. See https://bugs.eclipse.org/bugs/show_bug.cgi?id=563805.